# Switch from HashMap -> FxHashMap for better compilation performance. 