--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,6 +18,7 @@ bincode = "2.0.0-rc.3"
 # Fragile dependencies, minor updates often break the code
 hf-hub = "=0.3.2"
 tokenizers = { version = "=0.20.3", features = ["http"] }
+rustc-hash = "2.1.0"
 
 [features]
 python-bindings = ["pyo3"]