Rust's default HashMap Hash functions ( from SipHash ) are very slow since they do things like collision avoidance, and try to ensure uniform key distribution. These features seems un-needed, and for a library such as outlines-core focused on performance, it seems to be detrimental. Replacing the default rust HashMap with the hash map from `rustc_hash` makes this way faster, since it massively reduces hashing overhead. 

Benchmarks Before & After:
```
| Change   | Before [617a2c8e] <0.1.19>   | After [175ea196] <main~1>   | Ratio   | Benchmark (Parameter)                                                                                              |
|----------|------------------------------|-----------------------------|---------|--------------------------------------------------------------------------------------------------------------------|
| -        | 2.57±0.5s                    | 1.48±0.1s                   | 0.58    | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_fsm('complex_schema')                                    |
| -        | 1.35±0.5s                    | 786±70ms                    | 0.58    | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_fsm('simple_schema')                                     |
|          | 16.3±2μs                     | 15.6±0.3μs                  | 0.96    | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_regex('complex_schema')                                  |
|          | 6.10±2μs                     | 6.47±0.2μs                  | 1.06    | bench_json_schema.JsonSchemaBenchmark.time_json_schema_to_regex('simple_schema')                                   |
|          | 405M                         | 407M                        | 1.00    | bench_regex_guide.MemoryRegexGuideBenchmark.peakmem_regex_to_guide('complex_span_constrained_relation_extraction') |
|          | 401M                         | 403M                        | 1.01    | bench_regex_guide.MemoryRegexGuideBenchmark.peakmem_regex_to_guide('simple_phone')                                 |
| -        | 720±100ms                    | 351±10ms                    | 0.49    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('complex_phone')                                         |
| -        | 3.30±0.09s                   | 1.96±0.2s                   | 0.60    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('complex_span_constrained_relation_extraction')          |
| -        | 528±90ms                     | 279±20ms                    | 0.53    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('date')                                                  |
| -        | 387±50ms                     | 249±10ms                    | 0.64    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('email')                                                 |
|          | 381±60ms                     | 278±40ms                    | ~0.73   | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('ip')                                                    |
| -        | 347±70ms                     | 216±6ms                     | 0.62    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('simple_phone')                                          |
| -        | 306±40ms                     | 207±9ms                     | 0.68    | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('ssn')                                                   |
|          | 292±60ms                     | 253±30ms                    | ~0.87   | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('time')                                                  |
|          | 479±80ms                     | 303±10ms                    | ~0.63   | bench_regex_guide.RegexGuideBenchmark.time_regex_to_guide('url')                                                   |
```

`pyo3` already has built in conversion traits for `FxHashMap`s from `rustc_hash`, so python bindings / python usage does not change at all. All pre-commit hooks and tests pass. 
This adds the dependency of `rustc_hash`, and the `hashbrown` feature of `pyo3` for the conversion traits.